requirements.txt
Flask
Flask-SQLAlchemy
jinja2
Werkzeug
python-dotenv